+++ ugc/ugc.pro
@@ -10,10 +10,16 @@ include($$ROOT_DIR/common.pri)
 
 HEADERS += \
     api.hpp \
+    binary/header_v0.hpp \
+    binary/index_ugc.hpp \
+    binary/serdes.hpp \
+    binary/ugc_holder.hpp \
+    binary/visitors.hpp \
     serdes.hpp \
     storage.hpp \
     types.hpp \
 
 SOURCES += \
     api.cpp \
+    binary/serdes.cpp \
     storage.cpp \
